make base exchange service to keep things DRY
