+++ api/services/poloniex.service.js
@@ -1,31 +1,17 @@
 const config = require('../util/config');
-const logger = require('../util/logger');
 const NodeCache = require('node-cache');
-const _ = require('lodash');
 const Poloniex = require('poloniex-api-node');
+const BaseExchangeService = require('./base.exchange.service');
 
-class PoloniexService {
+class PoloniexService extends BaseExchangeService {
 
     constructor() {
+        super('Poloniex', new NodeCache({stdTTL: config.cacheTTL, checkperiod: 10}));
         this.poloniexApi = new Poloniex();
-        this.poloniexCache = new NodeCache({stdTTL: config.cacheTTL, checkperiod: 10});
-
-        this.poloniexCache.on('expired', async (tradingPairKey, value) => {
-            logger.info(`${tradingPairKey} expired updating poloniex order book`);
-            await this.updateOrderBook(tradingPairKey);
-        });
-    }
-
-    async connect(tradingPairs) {
-        this.tradingPairs = tradingPairs;
-
-        _.each(this.tradingPairs, async (tradingPair) => {
-            logger.info(`Initializing ${tradingPair} poloniex order book`);
-            await this.updateOrderBook(tradingPair);
-        });
     }
 
     async updateOrderBook(tradingPair) {
+        this.logger.debug(`${this.exchangeName} calling REST API for ${tradingPair}`);
         const orderBook = await this.poloniexApi.returnOrderBook(tradingPair.replace('-', '_'));
         const asks = orderBook.asks.map((ask) => {
             const pricePoint = Number(ask[0]);
@@ -45,18 +31,11 @@ class PoloniexService {
             }
         });
 
-        this.poloniexCache.set(tradingPair, {
+        this.exchangeCache.set(tradingPair, {
             asks: asks,
             bids: bids
         });
-    }
 
-    async getOrderBook(tradingPair) {
-        return this.poloniexCache.get(tradingPair);
-    }
-
-    // visible for testing
-    _setClient(poloniexClient) {
 
     }
 }
